Backend/requirements.txt
# requirements.txt
fastapi>=0.105.0
uvicorn>=0.24.0
pydantic>=2.5.2
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.25.2
openai>=1.3.7
replicate>=0.20.0
python-multipart>=0.0.6
sqlalchemy>=2.0.23
databases>=0.8.0
aiosqlite>=0.19.0
asyncpg>=0.29.0
jinja2>=3.1.2
pillow>=10.1.0